esesa-instagram-posts-to-blob-storage/requirements.txt
azure-common==1.1.28
azure-core==1.33.0
azure-nspkg==3.0.2
azure-storage==0.35.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
isodate==0.7.2
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.13.1
urllib3==2.3.0
